actualizacion y validacion de ingresos
