+++ src/main/java/Library.java
@@ -1,7 +1,4 @@
 import javax.swing.*;
-import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
 
 public class Library {
     public static void main(String[] args) {
@@ -12,230 +9,3 @@ public class Library {
     }
 }
 
-class LibraryGUI {
-    private JFrame frame;
-    private JPanel panel;
-    private JMenuBar menuBar;
-    private JMenu fileMenu;
-    private JMenu viewMenu;
-    private JMenu userMenu;
-    private JMenuItem exitItem;
-    private JMenuItem searchItem;
-    private JMenuItem insertItem;
-    private JMenuItem updateItem;
-    private JMenuItem insertUserItem;
-    private JMenuItem updateUserItem;
-    private JMenuItem deleteUserItem;
-    private CardLayout cardLayout;
-    private JPanel cards;
-    private JPanel searchPanel;
-    private JPanel insertPanel;
-    private JPanel updatePanel;
-    private JButton searchButton;
-    private JTextField searchField;
-    private JButton insertButton;
-    private JTextField titleField;
-    private JTextField authorField;
-    private JTextField genreField;
-    private JTextField yearField;
-    private LibraryDatabase database;
-
-    public LibraryGUI() {
-        frame = new JFrame("Work-up libreria");
-        frame.setSize(600, 400);
-        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-
-        panel = new JPanel();
-        panel.setLayout(new GridLayout(1, 1));
-
-        cardLayout = new CardLayout();
-        cards = new JPanel(cardLayout);
-
-        searchPanel = new JPanel();
-        searchPanel.setLayout(new FlowLayout());
-        searchButton = new JButton("Buscar libro");
-        searchField = new JTextField(20);
-        searchPanel.add(searchField);
-        searchPanel.add(searchButton);
-
-        updatePanel = new JPanel();
-        updatePanel.setLayout(new GridLayout( 2, 2));
-
-
-        insertPanel = new JPanel();
-        insertPanel.setLayout(new GridLayout(5, 6));
-        titleField = new JTextField(1);
-        authorField = new JTextField(1);
-        genreField = new JTextField(1);
-        yearField = new JTextField(1);
-        insertButton = new JButton("Insertar libro");
-        insertPanel.add(new JLabel("Título: "));
-        insertPanel.add(titleField);
-        insertPanel.add(new JLabel("Autor: "));
-        insertPanel.add(authorField);
-        insertPanel.add(new JLabel("Género: "));
-        insertPanel.add(genreField);
-        insertPanel.add(new JLabel("Año: "));
-        insertPanel.add(yearField);
-        insertPanel.add(insertButton);
-
-        cards.add(searchPanel, "Buscar libro");
-        cards.add(insertPanel, "Insertar libro");
-        cards.add(updatePanel, "Actualizar libro");
-
-
-        panel.add(cards);
-        frame.add(panel);
-
-        database = new LibraryDatabase();
-
-        searchButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                // Implementar la lógica de búsqueda de libros aquí
-                JOptionPane.showMessageDialog(frame, "Función de búsqueda aún no implementada");
-            }
-        });
-
-        insertButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                String title = titleField.getText();
-                String author = authorField.getText();
-                String genre = genreField.getText();
-                int year = Integer.parseInt(yearField.getText());
-
-                Book newBook = new Book(title, author, genre, year);
-                database.insertBook(newBook);
-
-                // Limpiar los campos de entrada después de la inserción
-                titleField.setText("");
-                authorField.setText("");
-                genreField.setText("");
-                yearField.setText("");
-
-                // Proporcionar retroalimentación al usuario, por ejemplo, mostrar un diálogo de mensaje
-                JOptionPane.showMessageDialog(frame, "Libro insertado exitosamente");
-            }
-        });
-
-        //menu de archivo
-        menuBar = new JMenuBar();
-        fileMenu = new JMenu("Archivo");
-        exitItem = new JMenuItem("Salir");
-
-        //menu de usuario
-        userMenu = new JMenu("Usuario");
-        insertUserItem = new JMenuItem("Insertar usuario");
-        updateUserItem = new JMenuItem("Actualizar usuario");
-        deleteUserItem = new JMenuItem("Eliminar usuario");
-
-        //menu de ver
-        viewMenu = new JMenu("Libros");
-        searchItem = new JMenuItem("Buscar libro");
-        insertItem = new JMenuItem("Insertar libro");
-        updateItem = new JMenuItem("actualizar libro");
-
-        fileMenu.add(exitItem);
-        viewMenu.add(searchItem);
-        viewMenu.add(insertItem);
-        viewMenu.add(updateItem);
-
-        userMenu.add(insertUserItem);
-        userMenu.add(updateUserItem);
-        userMenu.add(deleteUserItem);
-
-        menuBar.add(fileMenu);
-        menuBar.add(userMenu);
-        menuBar.add(viewMenu);
-        frame.setJMenuBar(menuBar);
-
-        searchItem.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                cardLayout.show(cards, "Buscar libro");
-            }
-        });
-
-        insertItem.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                cardLayout.show(cards, "Insertar libro");
-            }
-        });
-    }
-
-    public void showGUI() {
-        frame.setVisible(true);
-    }
-}
-
-class Book {
-    private String title;
-    private String author;
-    private String genre;
-    private int year;
-
-    public Book(String title, String author, String genre, int year) {
-        this.title = title;
-        this.author = author;
-        this.genre = genre;
-        this.year = year;
-    }
-
-    // Getters and setters
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public void setAuthor(String author) {
-        this.author = author;
-    }
-
-    public String getGenre() {
-        return genre;
-    }
-
-    public void setGenre(String genre) {
-        this.genre = genre;
-    }
-
-    public int getYear() {
-        return year;
-    }
-
-    public void setYear(int year) {
-        this.year = year;
-    }
-}
-
-class LibraryDatabase {
-    private Book[] books;
-    private int bookCount;
-
-    public LibraryDatabase() {
-        books = new Book[100]; // Suponiendo un máximo de 100 libros inicialmente
-        bookCount = 0;
-    }
-
-    public void insertBook(Book book) {
-        if (bookCount < books.length) {
-            books[bookCount] = book;
-            bookCount++;
-        } else {
-            // Manejar el caso cuando el arreglo está lleno, por ejemplo, redimensionar el arreglo
-        }
-    }
-
-    public Book[] getBooks() {
-        return books;
-    }
-}
